added ability to train in batches (and specifically mnist) 1) added reading of mnist data files into a vector of vectors of matrices ready for the nn 2) added functionality to train the nn in batches of size other than one by adding an average_gradients function to average the weight and bias gradients for each example in the batch
Note: This code runs but its efficacy for training has not. Due to below this will not output meaningful values anyway.
Note: The NN at this stage is not set up for the mnist problem. It is just outputting a value (that has been sigmoided) so in the -1 to 1 range. TODO add softmax end layer with 10 buckets to make into a classification NN.